Moved GOBitmap, GODC and GOFont to the gui/primitives subdirectory
(cherry picked from commit ad0a4602afb510cafdf8388ab7882ba0e30da500)
