+++ src/grandorgue/gui/GOGUIButton.cpp
@@ -11,8 +11,8 @@
 
 #include "config/GOConfigReader.h"
 #include "control/GOButtonControl.h"
+#include "primitives/GODC.h"
 
-#include "GODC.h"
 #include "GOGUIDisplayMetrics.h"
 #include "GOGUILayoutEngine.h"
 #include "GOGUIMouseState.h"
