+++ src/grandorgue/gui/GOGUIEnclosure.h
@@ -12,8 +12,9 @@
 
 #include <vector>
 
-#include "GOBitmap.h"
-#include "GOFont.h"
+#include "primitives/GOBitmap.h"
+#include "primitives/GOFont.h"
+
 #include "GOGUIControl.h"
 
 class GOEnclosure;
