+++ src/grandorgue/gui/GOGUIHW1Background.h
@@ -10,7 +10,8 @@
 
 #include <vector>
 
-#include "GOBitmap.h"
+#include "primitives/GOBitmap.h"
+
 #include "GOGUIControl.h"
 
 class GOGUIHW1Background : public GOGUIControl {
