+++ src/grandorgue/gui/GOGUIManual.cpp
@@ -12,8 +12,8 @@
 
 #include "config/GOConfigReader.h"
 #include "model/GOManual.h"
+#include "primitives/GODC.h"
 
-#include "GODC.h"
 #include "GOGUIDisplayMetrics.h"
 #include "GOGUILayoutEngine.h"
 #include "GOGUIMouseState.h"
