+++ src/grandorgue/gui/GOGUIManualBackground.cpp
@@ -9,8 +9,8 @@
 
 #include "config/GOConfigReader.h"
 #include "model/GOManual.h"
+#include "primitives/GODC.h"
 
-#include "GODC.h"
 #include "GOGUIDisplayMetrics.h"
 #include "GOGUILayoutEngine.h"
 #include "GOGUIPanel.h"
