+++ src/grandorgue/gui/primitives/GOBitmap.cpp
