+++ src/grandorgue/gui/primitives/GOBitmap.h
